Started on ingame UI

+++ src/main/abstractgame/world/entity/playermodules/Slider.java
@@ -1,21 +1,21 @@
 package abstractgame.world.entity.playermodules;
 
 public class Slider extends Customization {
-	double value;
+	float value;
 	
-	final double min;
-	final double max;
+	final float min;
+	final float max;
 	final int graduation;
 	
 	/** A graduation of 0 means that the slider is completely smooth */
-	public Slider(String name, String description, double min, double max, int graduation) {
+	public Slider(String name, String description, float min, float max, int graduation) {
 		super(name, description);
 		this.min = min;
 		this.max = max;
 		this.graduation = graduation;
 	}
 
-	public double getValue() {
+	public float getValue() {
 		return value;
 	}
 }
